Fix nanomsg frame sending
Nanomsg was not actually sending when it was called within the callback from AVFoundation. That context interferes with nanomsg somehow. Altered the code to use an inproc nanomsg queue to first transfer the jpeg frame to the main thread and then send out to tcp via nanomsg from there. That works.

Also fixed a small bit of uclop; it wasn't handling flags appropriately.