Fix device use of devices with dash in UDID
IOS devices with a dash in UDID were still not working even after recent changes.
This fix should address the issue. This change has in fact been tested against such a device and is working.

Change was to use the newer API from Apple that was being used to list devices to also select the device.